+++ packages/frontend/src/components/pages/index.ts
@@ -1,3 +1,0 @@
-export { HomePage } from "./Home.page";
-export { AppsPage } from "./Apps.page";
-export { AboutPage } from "./About.page";