+++ packages/frontend/src/components/routes/Apps.route.ts
@@ -1,1 +1,0 @@
-export { AppsPage as AppsRoute } from "../pages/Apps.page";