+++ packages/frontend/src/data/common.ts
@@ -1,0 +1,14 @@
+import favicon from "../assets/icon.png";
+import background from "../assets/bg.jpg";
+
+const common = {
+  title: "Learn Chinese Platform",
+  author: "Felix Pojtinger",
+  site: "https://learn-chinese.tk",
+  language: "en",
+  color: "#e31818",
+  favicon,
+  background
+};
+
+export { common };