+++ packages/frontend/src/data/decks.ts
@@ -1,0 +1,17 @@
+import { common } from "./common";
+
+const decks = {
+  head: {
+    title: "Decks",
+    description: "Manage decks and export them as an Anki packages or PDFs.",
+    author: common.author,
+    keywords: ["decks", "anki", "list", "vocab"],
+    favicon: common.favicon,
+    siteTitle: common.title,
+    themeColor: common.color,
+    lang: common.language
+  },
+  background: common.background
+};
+
+export { decks };