+++ packages/frontend/src/data/dictionary.ts
@@ -1,0 +1,17 @@
+import { common } from "./common";
+
+const dictionary = {
+  head: {
+    title: "Dictionary",
+    description: "Search for Hanzi, Pinyin or English.",
+    author: common.author,
+    keywords: ["home", "overview"],
+    favicon: common.favicon,
+    siteTitle: common.title,
+    themeColor: common.color,
+    lang: common.language
+  },
+  background: common.background
+};
+
+export { dictionary };