+++ packages/frontend/src/data/notifications.ts
@@ -1,0 +1,17 @@
+import { common } from "./common";
+
+const notifications = {
+  head: {
+    title: "Notifications",
+    description: "Take a look at your notifications.",
+    author: common.author,
+    keywords: ["notifications", "updates"],
+    favicon: common.favicon,
+    siteTitle: common.title,
+    themeColor: common.color,
+    lang: common.language
+  },
+  background: common.background
+};
+
+export { notifications };