+++ packages/frontend/src/data/settings.ts
@@ -1,0 +1,17 @@
+import { common } from "./common";
+
+const settings = {
+  head: {
+    title: "Settings",
+    description: "Change the settings of this app.",
+    author: common.author,
+    keywords: ["settings", "preferences"],
+    favicon: common.favicon,
+    siteTitle: common.title,
+    themeColor: common.color,
+    lang: common.language
+  },
+  background: common.background
+};
+
+export { settings };