+++ packages/frontend/src/types/page.type.ts
@@ -1,5 +1,0 @@
-interface IPage {
-  props?: any;
-}
-
-export { IPage };